Updated programs to need a main function
I've updated a lot of commands to be more modern and consistently
written (a lot of 2.x default programs were written in ways that are, in
hindsight, esoteric, because the APIs were less defined and I hadn't
ironed out Pythinux's kinks). I've also added a main() function to allow
for libraries to be imported without any text being displayed by passing
`libMode=True` to `load_program()`. It really is quite that simple.

I've also updated most programs to work through an API.