Removed `logoff` 'command'
In Pythinux 2.x, `logoff` was actually a command handled by `terminal`.
Because it wasn't a real program, help didn't list it. As such, a janky
solution to just return the program list plus "logoff" as the program
list.

When `terminal` became replaced with the MUCH better `shell`, `logoff`
was not implemented. However, help.help() still returned it, because the
old code did that, and I didn't notice. Until now.

`logoff` no longer exists, so I've removed it from help.help().