Deprecated giveOutput()
giveOutput() now raises a DeprecationWarning that states that the
function will be removed with release 3.1. The reason for this is
because its need has come and gone; it was never reliable in the first
place, and now that the main() function has been implemented, programs
can be both libraries and frontend programs.