"pkm clear" now removes ALL packages
Now, the clear all programs loop runs until no programs are installed,
which is a rather brute-force manner of resolving dependencies, althogh
it is effective, as it runs through the program list multiple times to
ensure no programs remain.

However, in the case of a circular dependency (where foo depends on bar
and bar depends on foo), the program will run forever. Thankfully, it is
impossible to reach this state by normal means, as installing either foo
or bar will also cause an infinite installation loop, as foo installs
bar before it installs itself, and bar does the same. Therefore, it's a
non-issue.