Added `pkm dist-upgrade` and Pythinux Upgrade Framework
The `pythinux dist-upgrade` command allows you to install new Pythinux
versions in-place. This works by downloading a remote ZIP file and
unzipping it in the directory ABOVE the root directory, which allows for
kernel upgrades. This leads to an incremental upgrade as well, since
only the necessary changes are copied in.