Added Disallowed Contributions clause
In case you're angry at the clause, I have nothing against making a contribution and using it as proof of one's skills. However, if you have no intention of being useful to the project, and the only thing you want to do is add a low-quality spam commit to claim you've worked on Pythinux, you are **not** welcome in our community until you correct your attitude.