Added shell.pipeCommandOutput (#27)
This function calls takes three args - a user and two commands to run.
It uses shell.getCommandOutput() on the first command and passes that
into the second command's pipe() function. This has been tested, and it
works.

As of now, no commands support piping and the shell itself cannot pipe
yet, but that's soon to change.

Signed-off-by: Gordinator <gordinator@gordinator.org>