Added debug.grant_debugging(), debug.revoke_debugging(), debug.cleanup()
These functions can be used to add and remove debuggers. Soon, they will
ask for user permission interactively.

The debug.cleanup() function is called by shell.init() and also debugmgr
and essentially makes sure invalid programs are not allowed on the
debugger list.

The debug.system object is a list of system debuggers.

Signed-off-by: Gordinator <gordinator@gordinator.org>